Moving more folders
added  components/applyPage folder
- moved in form folder

added community/leaderboard folder
- moved in achievement-card and badge-modal folder

added components used multiple times in the website to constants folder
