+++ src/app/profile/page.tsx
@@ -3,7 +3,7 @@ import { Sidebar } from "@/component/components";
 import React from "react";
 import { Col, Container, Row } from "react-bootstrap";
 import styles from "./profile.module.css";
-import AchievementCard from "@/component/components/achievement-card";
+import AchievementCard from "@/component/components/communityProgram/leaderboard/achievement-card";
 
 const Profile = () => {
   const users = Array.from(Array(10).keys());
