+++ src/components/applyPage/form/formHeader.css
@@ -1,5 +1,5 @@
 .bg-dot {
-    background-image: url("../../assets/images/dots.svg"); 
+    background-image: url("../../../assets/images/dots.svg"); 
     background-repeat: no-repeat;
     background-position:right;
     z-index:1;
