+++ src/components/applyPage/form/index.tsx
@@ -1,3 +1,4 @@
+"use client"
 import React, { useState, useEffect } from "react";
 import InputField from "./inputField";
 import FormHeader from "./formHeader";
