+++ src/components/index.tsx
@@ -1,20 +1,12 @@
 "use client"
-export * from './c4gt23/mentoringProgram/productList/productList'
-export * from './header/header';
-export * from './sidebar';
-export * from './achievement-card';
-export * from './badge-modal';
-export * from './c4gt23/mentoringProgram/productCard/productCard';
-export * from './c4gt23/mentoringProgram/searchBar/search';
-export * from './footer';
-export * from './navbar';
-export * from './home/homeHeader';
-export * from './home/homeBody';
+
+//Home Page Components
+export * from './homePage/homeHeader';
+export * from './homePage/homeBody';
 
 // C4GT 2022 page components
 export * from './c4gt22/headerCard';
 export * from './c4gt22/infoBar';
-export * from './c4gt22/infoBar/infoIcon';
 export * from './c4gt22/infoCard';
 export * from './c4gt22/projectCard';
 export * from './c4gt22/projectCard/projectList';
@@ -23,6 +15,20 @@ export * from './c4gt22/mentorCard';
 
 // C4GT 2023 page components
 export * from './c4gt23/infoBar23';
+export * from './c4gt23/mentoringProgram/productList/productList'
+export * from './c4gt23/mentoringProgram/productCard/productCard';
+export * from './c4gt23/mentoringProgram/searchBar/search';
+
+// Community Program Components
+export * from './communityProgram/leaderboard/achievement-card';
+export * from './communityProgram/leaderboard/badge-modal';
 
 //Form page 
-export * from './form';
+export * from './applyPage/form';
+
+// Constant Components
+export * from '../constants/header/header';
+export * from '../constants/sidebar';
+export * from '../constants/infoIcon';
+export * from '../constants/footer';
+export * from '../constants/navbar';
